Add features for process admins
